refactor(executor): split join_scan.rs by join algorithm
Split the monolithic join_scan.rs (2,149 lines) into a module directory
with focused submodules:

- `mod.rs`: Public API, join execution orchestration, NATURAL/USING join
  condition generation, and ON clause validation
- `predicates.rs`: Predicate helper functions (table name extraction,
  predicate reference analysis)
- `pushdown.rs`: Right-table predicate pushdown for SEMI/ANTI and OUTER
  joins
- `semi_anti.rs`: SEMI/ANTI join optimization logic (INL cost decision,
  condition parsing)
- `inl.rs`: Index Nested Loop implementation (exact-match and prefix-scan
  strategies)

Benefits:
- Smaller, focused files (~400-500 lines each instead of 2,149)
- Join algorithms isolated for easier optimization
- Predicate handling separated from execution
- Clearer code paths for different join types
- No public API changes

Closes #4849
