+++ crates/vibesql-executor/src/select/scan/join_scan/mod.rs
@@ -28,20 +28,19 @@
 
 use std::collections::{HashMap, HashSet};
 
-use ahash::AHashSet;
-
 use crate::{
     errors::ExecutorError,
-    evaluator::CombinedExpressionEvaluator,
-    optimizer::{
-        combine_with_and,
-        where_pushdown::{extract_referenced_tables_branch, flatten_conjuncts},
-        PredicatePlan,
-    },
+    optimizer::PredicatePlan,
     schema::CombinedSchema,
     select::cte::CteResult,
     timeout::TimeoutContext,
 };
+
+// Submodules
+mod inl;
+mod predicates;
+mod pushdown;
+mod semi_anti;
 
 /// Base threshold for using Index Nested Loop semi-join instead of hash join.
 /// If left side has fewer rows than this, INL is always preferred.
@@ -86,7 +85,7 @@ where
         vibesql_ast::JoinType::RightOuter | vibesql_ast::JoinType::FullOuter => {
             // For RIGHT/FULL OUTER JOIN: Remove predicates that reference ONLY the left table
             if let Some(where_expr) = where_clause {
-                filter_out_nullable_side_predicates(left, where_expr, database)
+                pushdown::filter_out_nullable_side_predicates(left, where_expr, database)
             } else {
                 None
             }
@@ -120,12 +119,12 @@ where
         if left_row_count > 0 {
             // Determine if INL should be used based on cost estimation
             let should_use_inl =
-                should_use_inl_for_semi_join(left_row_count, right, condition, database);
+                semi_anti::should_use_inl_for_semi_join(left_row_count, right, condition, database);
 
             if should_use_inl {
                 // Try exact-match INL first (requires full PK coverage)
                 if let Some(result) =
-                    try_index_nested_loop_semi_join(&left_result, right, condition, database)?
+                    inl::try_index_nested_loop_semi_join(&left_result, right, condition, database)?
                 {
                     return Ok(result);
                 }
@@ -134,7 +133,7 @@ where
                 // This handles cases like TPC-H Q4 where we join on L_ORDERKEY but the
                 // index is on (L_ORDERKEY, L_LINENUMBER) - we can still use a prefix scan
                 if let Some(result) =
-                    try_prefix_scan_semi_join(&left_result, right, condition, database)?
+                    inl::try_prefix_scan_semi_join(&left_result, right, condition, database)?
                 {
                     return Ok(result);
                 }
@@ -165,7 +164,7 @@ where
         vibesql_ast::JoinType::Semi | vibesql_ast::JoinType::Anti => {
             // Extract right-table-only predicates from the join condition
             if let Some(cond) = condition {
-                extract_right_only_predicates(right, cond, database)
+                pushdown::extract_right_only_predicates(right, cond, database)
             } else {
                 None
             }
@@ -174,7 +173,7 @@ where
             // For LEFT/FULL OUTER JOIN: Remove predicates that reference ONLY the right table
             // These must be evaluated post-join when NULL values from unmatched rows are present
             if let Some(where_expr) = where_clause {
-                filter_out_nullable_side_predicates(right, where_expr, database)
+                pushdown::filter_out_nullable_side_predicates(right, where_expr, database)
             } else {
                 None
             }
@@ -673,915 +672,6 @@ fn generate_natural_join_condition(
     }
 
     Ok(condition)
-}
-
-/// Extract right-table-only predicates from a join condition for SEMI/ANTI join pushdown
-///
-/// For IN subquery to SEMI JOIN conversions, the subquery's WHERE clause predicates are
-/// combined into the JOIN condition. This function extracts predicates that only reference
-/// the right-side table(s) so they can be pushed down to the right-side scan for index selection.
-///
-/// Example: For `WHERE ol_i_id IN (SELECT s_i_id FROM stock WHERE s_w_id = ? AND s_quantity < ?)`
-/// The JOIN condition is: `ol_i_id = s_i_id AND s_w_id = ? AND s_quantity < ?`
-/// This function extracts: `s_w_id = ? AND s_quantity < ?` (right-only predicates)
-///
-/// Returns None if no right-only predicates can be extracted.
-fn extract_right_only_predicates(
-    right_from: &vibesql_ast::FromClause,
-    condition: &vibesql_ast::Expression,
-    database: &vibesql_storage::Database,
-) -> Option<vibesql_ast::Expression> {
-    // Get the table name(s) from the right-side FROM clause
-    let right_tables = extract_table_names_from_from_clause(right_from);
-    if right_tables.is_empty() {
-        return None;
-    }
-
-    // We need to build a minimal schema to analyze predicates
-    // For each table, get its schema from the database
-    let mut right_table_set: HashSet<String> = HashSet::new();
-    let mut schema_builder = crate::schema::SchemaBuilder::new();
-
-    for table_name in &right_tables {
-        // Normalize table name for lookup
-        let normalized = table_name.to_lowercase();
-        if let Some(table) = database.get_table(&normalized) {
-            // Add table to schema builder (duplicates are tracked, not rejected)
-            schema_builder.add_table(table_name.clone(), table.schema.clone());
-            right_table_set.insert(table_name.clone());
-            // Also add normalized version
-            right_table_set.insert(normalized);
-        }
-    }
-
-    if right_table_set.is_empty() {
-        return None;
-    }
-
-    let right_schema = schema_builder.build();
-
-    // Flatten the condition into conjuncts (AND-separated predicates)
-    let conjuncts = flatten_conjuncts(condition);
-
-    // Filter to keep only predicates that reference only right-side tables
-    let right_only_predicates: Vec<vibesql_ast::Expression> = conjuncts
-        .into_iter()
-        .filter(|pred| {
-            // extract_referenced_tables_branch returns Option<HashSet<String>>
-            // None means the expression couldn't be analyzed (skip it)
-            // Some(empty set) means no tables referenced (skip it)
-            // Some(non-empty set) - check if all tables are in right_table_set
-            match extract_referenced_tables_branch(pred, &right_schema) {
-                Some(ref tables) if !tables.is_empty() => tables.iter().all(|t| {
-                    let t_lower = t.to_lowercase();
-                    right_table_set.contains(t) || right_table_set.contains(&t_lower)
-                }),
-                _ => false,
-            }
-        })
-        .collect();
-
-    if right_only_predicates.is_empty() {
-        return None;
-    }
-
-    // Combine predicates with AND
-    combine_with_and(right_only_predicates)
-}
-
-/// Filter out predicates that reference ONLY the nullable side of an outer join.
-///
-/// For LEFT/FULL OUTER JOIN, the right side is "nullable" - it can produce NULL values
-/// for unmatched rows. Predicates like `right.col IS NULL` test for these join-produced NULLs,
-/// not for NULL values stored in the table. If we push such predicates to the right-side scan,
-/// they filter on stored NULLs instead of join-produced NULLs, causing incorrect results.
-///
-/// This function takes a WHERE clause and returns a modified version with nullable-side-only
-/// predicates removed. These predicates will be evaluated post-join instead.
-///
-/// Example: For `LEFT JOIN t2 ON t1.id = t2.tid WHERE t2.tid IS NULL`
-/// - Input: `t2.tid IS NULL`
-/// - Output: None (the entire predicate references only the nullable side)
-///
-/// Example: For `LEFT JOIN t2 ON t1.id = t2.tid WHERE t1.id > 5 AND t2.tid IS NULL`
-/// - Input: `t1.id > 5 AND t2.tid IS NULL`
-/// - Output: `t1.id > 5` (keep left-side predicate, remove right-side predicate)
-fn filter_out_nullable_side_predicates(
-    nullable_side_from: &vibesql_ast::FromClause,
-    where_expr: &vibesql_ast::Expression,
-    database: &vibesql_storage::Database,
-) -> Option<vibesql_ast::Expression> {
-    // Get the table name(s) from the nullable-side FROM clause
-    let nullable_tables = extract_table_names_from_from_clause(nullable_side_from);
-    if nullable_tables.is_empty() {
-        // No tables to filter, return original expression
-        return Some(where_expr.clone());
-    }
-
-    // Build a schema and table set for the nullable side
-    let mut nullable_table_set: HashSet<String> = HashSet::new();
-    let mut schema_builder = crate::schema::SchemaBuilder::new();
-
-    for table_name in &nullable_tables {
-        // Normalize table name for lookup
-        let normalized = table_name.to_lowercase();
-        if let Some(table) = database.get_table(&normalized) {
-            // Add table to schema builder (duplicates are tracked, not rejected)
-            schema_builder.add_table(table_name.clone(), table.schema.clone());
-            nullable_table_set.insert(table_name.clone());
-            // Also add normalized version for case-insensitive matching
-            nullable_table_set.insert(normalized);
-        } else {
-            // Table not found in database - might be a subquery alias
-            // Still add to the set so we can check column references
-            nullable_table_set.insert(table_name.clone());
-            nullable_table_set.insert(normalized);
-        }
-    }
-
-    let nullable_schema = schema_builder.build();
-
-    // Flatten the WHERE clause into conjuncts (AND-separated predicates)
-    let conjuncts = flatten_conjuncts(where_expr);
-
-    // Filter to KEEP predicates that do NOT reference only nullable-side tables
-    // (i.e., remove predicates that reference ONLY the nullable side)
-    let kept_predicates: Vec<vibesql_ast::Expression> = conjuncts
-        .into_iter()
-        .filter(|pred| {
-            // extract_referenced_tables_branch returns Option<HashSet<String>>
-            // None means the expression couldn't be analyzed (keep it, evaluate post-join)
-            // Some(empty set) means no tables referenced (keep it - e.g., constant expression)
-            // Some(non-empty set) - check if ALL tables are in nullable_table_set
-            //   If ALL are nullable-side-only -> REMOVE (don't keep)
-            //   If ANY are from other tables -> KEEP
-            match extract_referenced_tables_branch(pred, &nullable_schema) {
-                Some(ref tables) if !tables.is_empty() => {
-                    // Check if ALL referenced tables are in the nullable set
-                    let all_nullable = tables.iter().all(|t| {
-                        let t_lower = t.to_lowercase();
-                        nullable_table_set.contains(t) || nullable_table_set.contains(&t_lower)
-                    });
-                    // KEEP if NOT all-nullable (i.e., at least one non-nullable table referenced)
-                    !all_nullable
-                }
-                Some(_) => true, // Empty set - keep (no table refs, e.g., literals)
-                None => {
-                    // Couldn't analyze - check if predicate mentions nullable table columns
-                    // For safety, keep predicates we can't analyze and evaluate post-join
-                    // But first, do a simple check for column references
-                    !predicate_references_only_tables(pred, &nullable_table_set)
-                }
-            }
-        })
-        .collect();
-
-    // Combine remaining predicates with AND (returns None if empty)
-    combine_with_and(kept_predicates)
-}
-
-/// Simple check if a predicate references only tables from a given set.
-/// This is a fallback for when extract_referenced_tables_branch returns None.
-fn predicate_references_only_tables(
-    expr: &vibesql_ast::Expression,
-    table_set: &HashSet<String>,
-) -> bool {
-    match expr {
-        vibesql_ast::Expression::ColumnRef(col_id)
-            if col_id.schema_canonical().is_none() && col_id.table_canonical().is_some() =>
-        {
-            let t = col_id.table_canonical().unwrap();
-            let t_lower = t.to_lowercase();
-            table_set.contains(t) || table_set.contains(&t_lower)
-        }
-        vibesql_ast::Expression::ColumnRef(col_id)
-            if col_id.schema_canonical().is_none() && col_id.table_canonical().is_none() =>
-        {
-            // Unqualified column - can't determine which table, assume it might be from nullable
-            // side Return false to be conservative (keep the predicate)
-            false
-        }
-        vibesql_ast::Expression::ColumnRef(_) => {
-            // Schema-qualified or other cases, be conservative
-            false
-        }
-        vibesql_ast::Expression::IsNull { expr: inner, .. } => {
-            predicate_references_only_tables(inner, table_set)
-        }
-        vibesql_ast::Expression::UnaryOp { expr: inner, .. } => {
-            predicate_references_only_tables(inner, table_set)
-        }
-        vibesql_ast::Expression::BinaryOp { left, right, .. } => {
-            // For binary ops, check if BOTH sides reference only the table set
-            let left_only = predicate_references_only_tables(left, table_set);
-            let right_only = predicate_references_only_tables(right, table_set);
-            // If either side references other tables, this is not "only" nullable-side
-            left_only && right_only
-        }
-        vibesql_ast::Expression::Literal(_) => {
-            // Literals don't reference any table
-            true
-        }
-        // For other expression types, be conservative
-        _ => false,
-    }
-}
-
-/// Extract table names from a FROM clause
-fn extract_table_names_from_from_clause(from: &vibesql_ast::FromClause) -> Vec<String> {
-    let mut tables = Vec::new();
-    collect_table_names(from, &mut tables);
-    tables
-}
-
-fn collect_table_names(from: &vibesql_ast::FromClause, tables: &mut Vec<String>) {
-    match from {
-        vibesql_ast::FromClause::Table { name, alias, .. } => {
-            // Use alias if present, otherwise use table name
-            tables.push(alias.clone().unwrap_or_else(|| name.clone()));
-        }
-        vibesql_ast::FromClause::Join { left, right, .. } => {
-            collect_table_names(left, tables);
-            collect_table_names(right, tables);
-        }
-        vibesql_ast::FromClause::Subquery { alias, .. } => {
-            // Subquery alias is required (String, not Option<String>)
-            tables.push(alias.clone());
-        }
-        vibesql_ast::FromClause::Values { alias, .. } => {
-            // VALUES alias is required (String, not Option<String>)
-            tables.push(alias.clone());
-        }
-    }
-}
-
-/// Determine whether to use Index Nested Loop (INL) for a semi-join.
-///
-/// This implements a cost-based decision between INL and hash semi-join:
-///
-/// 1. **Always use INL** when left_rows < INL_BASE_THRESHOLD (1000)
-///    - Small left side means few index lookups, definitely faster than hash
-///
-/// 2. **Use INL when right table is much larger** (right_rows / left_rows > ratio threshold)
-///    - Hash join scans the entire right table to build hash table
-///    - INL does left_rows index lookups, each returning ~(right_rows/distinct_keys) rows
-///    - For semi-join, we stop at first match, so avg cost is even lower
-///
-/// 3. **Cap at INL_MAX_THRESHOLD** to prevent excessive random I/O
-///
-/// Example (TPC-H Q4):
-/// - left_rows = 5,406 (orders after date filter)
-/// - right_rows = 600,000 (lineitem)
-/// - ratio = 600,000 / 5,406 ≈ 111 >> 10
-/// - Decision: Use INL (index lookups) instead of hash (full lineitem scan)
-fn should_use_inl_for_semi_join(
-    left_row_count: usize,
-    right_from: &vibesql_ast::FromClause,
-    _condition: &Option<vibesql_ast::Expression>,
-    database: &vibesql_storage::Database,
-) -> bool {
-    let debug = std::env::var("INL_DECISION_DEBUG").is_ok();
-
-    // Rule 1: Always use INL for small left sides
-    if left_row_count < INL_BASE_THRESHOLD {
-        if debug {
-            eprintln!(
-                "[INL_DECISION] left_rows={} < base_threshold={}, using INL",
-                left_row_count, INL_BASE_THRESHOLD
-            );
-        }
-        return true;
-    }
-
-    // Rule 3: Never use INL for very large left sides (too much random I/O)
-    if left_row_count > INL_MAX_THRESHOLD {
-        if debug {
-            eprintln!(
-                "[INL_DECISION] left_rows={} > max_threshold={}, using hash join",
-                left_row_count, INL_MAX_THRESHOLD
-            );
-        }
-        return false;
-    }
-
-    // Rule 2: Use cost-based decision for medium-sized left sides
-    // Get right table cardinality
-    let right_row_count = match right_from {
-        vibesql_ast::FromClause::Table { name, .. } => {
-            database.get_table(name).map(|t| t.row_count()).unwrap_or(0)
-        }
-        _ => {
-            // Complex right side (subquery/join), can't estimate, use hash join
-            if debug {
-                eprintln!("[INL_DECISION] Complex right side, using hash join");
-            }
-            return false;
-        }
-    };
-
-    // Calculate size ratio
-    let ratio = if left_row_count > 0 { right_row_count / left_row_count } else { 0 };
-
-    // Decision: Use INL if right table is significantly larger than left
-    let use_inl = ratio >= INL_SIZE_RATIO_THRESHOLD;
-
-    if debug {
-        eprintln!(
-            "[INL_DECISION] left_rows={}, right_rows={}, ratio={}, threshold={}, decision={}",
-            left_row_count,
-            right_row_count,
-            ratio,
-            INL_SIZE_RATIO_THRESHOLD,
-            if use_inl { "INL" } else { "hash" }
-        );
-    }
-
-    use_inl
-}
-
-/// Try to execute a SEMI join using Index Nested Loop (INL) strategy.
-///
-/// This optimization is used when:
-/// 1. The left side is small (< INL_BASE_THRESHOLD rows)
-/// 2. The right side is a simple table (not a subquery or join)
-/// 3. There's an equi-join condition on a column with an index
-///
-/// For each distinct join key from the left side, we do point lookups on the
-/// right table instead of scanning all matching rows.
-///
-/// Returns Some(result) if INL was used successfully, None to fall back to hash join.
-fn try_index_nested_loop_semi_join(
-    left_result: &super::FromResult,
-    right_from: &vibesql_ast::FromClause,
-    condition: &Option<vibesql_ast::Expression>,
-    database: &vibesql_storage::Database,
-) -> Result<Option<super::FromResult>, ExecutorError> {
-    // Must have a join condition
-    let cond = match condition {
-        Some(c) => c,
-        None => return Ok(None),
-    };
-
-    // Right side must be a simple table (not a join or subquery)
-    let (right_table_name, _right_alias) = match right_from {
-        vibesql_ast::FromClause::Table { name, alias, .. } => (name.clone(), alias.clone()),
-        _ => return Ok(None), // Complex right side, can't use INL
-    };
-
-    // Get the right table
-    let right_table = match database.get_table(&right_table_name) {
-        Some(t) => t,
-        None => return Ok(None),
-    };
-
-    // Parse the join condition to extract:
-    // 1. The equi-join columns (left_col = right_col)
-    // 2. Additional filter predicates on the right table
-    let (equi_join, right_filters) =
-        match parse_semi_join_condition(cond, left_result, &right_table_name) {
-            Some(parsed) => parsed,
-            None => return Ok(None),
-        };
-
-    // Check if the right table has a usable index for point lookups
-    // We need an index that starts with the join key column
-    let pk_columns = match &right_table.schema.primary_key {
-        Some(cols) => cols.clone(),
-        None => return Ok(None), // No primary key, can't use INL
-    };
-
-    // For TPC-C Stock-Level: STOCK has PK (s_w_id, s_i_id)
-    // The join is on s_i_id, but we also have s_w_id = ? in the filter
-    // So we can build a composite key for point lookups
-
-    // Try to extract constant predicates from right_filters that can be combined with join key
-    let constant_prefix =
-        extract_constant_prefix_for_pk(&right_filters, &pk_columns, &equi_join.right_col);
-
-    // Build the index lookup key template
-    // For Stock-Level: key = [s_w_id (from filter), s_i_id (from join)]
-    let lookup_key_template =
-        match build_lookup_key_template(&pk_columns, &equi_join.right_col, &constant_prefix) {
-            Some(template) => template,
-            None => return Ok(None),
-        };
-
-    // Get the primary key index
-    let pk_index = match right_table.primary_key_index() {
-        Some(idx) => idx,
-        None => return Ok(None),
-    };
-
-    // Get left column index for extracting join keys
-    let left_col_idx = match find_column_index(&left_result.schema, &equi_join.left_col) {
-        Some(idx) => idx,
-        None => return Ok(None),
-    };
-
-    // Build evaluator for right-side filter (residual predicates after PK lookup)
-    let right_schema =
-        CombinedSchema::from_table(right_table_name.clone(), right_table.schema.clone());
-    let residual_filter = build_residual_filter(&right_filters, &pk_columns);
-    let evaluator = residual_filter
-        .as_ref()
-        .map(|_| CombinedExpressionEvaluator::with_database(&right_schema, database));
-
-    // Collect distinct join keys from left side
-    let left_slice = left_result.as_slice();
-    let mut seen_keys: AHashSet<vibesql_types::SqlValue> =
-        AHashSet::with_capacity(left_slice.len());
-    let mut matching_keys: AHashSet<vibesql_types::SqlValue> = AHashSet::new();
-
-    for left_row in left_slice {
-        let join_key = &left_row.values[left_col_idx];
-
-        // Skip NULL join keys
-        if *join_key == vibesql_types::SqlValue::Null {
-            continue;
-        }
-
-        // Skip duplicate keys (we only need one match for SEMI join)
-        if seen_keys.contains(join_key) {
-            continue;
-        }
-        seen_keys.insert(join_key.clone());
-
-        // Build the full lookup key
-        let lookup_key: Vec<vibesql_types::SqlValue> = lookup_key_template
-            .iter()
-            .map(|slot| match slot {
-                KeySlot::Constant(v) => v.clone(),
-                KeySlot::JoinKey => join_key.clone(),
-            })
-            .collect();
-
-        // Do point lookup
-        // Issue #3790: Use get_row() which returns None for deleted rows
-        if let Some(&row_idx) = pk_index.get(&lookup_key) {
-            let right_row = match right_table.get_row(row_idx) {
-                Some(row) => row,
-                None => continue, // Row deleted or invalid
-            };
-
-            // Apply residual filter if any
-            let passes = if let (Some(filter), Some(eval)) = (&residual_filter, &evaluator) {
-                eval.clear_cse_cache();
-                matches!(eval.eval(filter, right_row), Ok(vibesql_types::SqlValue::Boolean(true)))
-            } else {
-                true // No residual filter
-            };
-
-            if passes {
-                matching_keys.insert(join_key.clone());
-            }
-        }
-    }
-
-    // Build result: all left rows whose join key is in matching_keys
-    let result_rows: Vec<vibesql_storage::Row> = left_slice
-        .iter()
-        .filter(|row| {
-            let key = &row.values[left_col_idx];
-            matching_keys.contains(key)
-        })
-        .cloned()
-        .collect();
-
-    Ok(Some(super::FromResult::from_rows(left_result.schema.clone(), result_rows)))
-}
-
-/// Parsed equi-join condition
-struct EquiJoinInfo {
-    left_col: String,
-    right_col: String,
-}
-
-/// Parse a semi-join condition to extract equi-join and right-side filters.
-fn parse_semi_join_condition(
-    cond: &vibesql_ast::Expression,
-    left_result: &super::FromResult,
-    right_table_name: &str,
-) -> Option<(EquiJoinInfo, Option<vibesql_ast::Expression>)> {
-    let conjuncts = flatten_conjuncts(cond);
-
-    let left_tables: HashSet<String> =
-        left_result.schema.table_schemas.keys().map(|s| s.canonical().to_string()).collect();
-
-    let right_table_lower = right_table_name.to_lowercase();
-
-    let mut equi_join: Option<EquiJoinInfo> = None;
-    let mut right_only_preds: Vec<vibesql_ast::Expression> = Vec::new();
-
-    for pred in conjuncts {
-        // Check if this is an equi-join predicate (col1 = col2)
-        if let vibesql_ast::Expression::BinaryOp {
-            left,
-            op: vibesql_ast::BinaryOperator::Equal,
-            right,
-        } = &pred
-        {
-            if let (
-                vibesql_ast::Expression::ColumnRef(left_col_id),
-                vibesql_ast::Expression::ColumnRef(right_col_id),
-            ) = (left.as_ref(), right.as_ref())
-            {
-                if left_col_id.schema_canonical().is_some()
-                    || right_col_id.schema_canonical().is_some()
-                {
-                    // Schema-qualified, skip this predicate
-                    right_only_preds.push(pred);
-                    continue;
-                }
-                let left_tbl = left_col_id.table_canonical();
-                let right_tbl = right_col_id.table_canonical();
-                let left_col = left_col_id.column_canonical();
-                let right_col = right_col_id.column_canonical();
-
-                // Determine which column is from left and which from right
-                let left_tbl_lower = left_tbl.map(|s| s.to_lowercase());
-                let right_tbl_lower = right_tbl.map(|s| s.to_lowercase());
-
-                let left_col_lower = left_col.to_lowercase();
-
-                // Check if left_col is from left tables and right_col is from right table
-                // When table qualifier is None, check if column exists in any left table's schema
-                let left_is_left =
-                    left_tbl_lower.as_ref().map(|t| left_tables.contains(t)).unwrap_or(false)
-                        || left_result.schema.table_schemas.values().any(|(_, schema)| {
-                            schema.columns.iter().any(|c| c.name.to_lowercase() == left_col_lower)
-                        });
-                let right_is_right =
-                    right_tbl_lower.as_ref().map(|t| t == &right_table_lower).unwrap_or(true);
-
-                if left_is_left && right_is_right && equi_join.is_none() {
-                    equi_join = Some(EquiJoinInfo {
-                        left_col: left_col.to_string(),
-                        right_col: right_col.to_string(),
-                    });
-                    continue;
-                }
-
-                // Check the reverse: right_col from left, left_col from right
-                let right_is_left =
-                    right_tbl_lower.as_ref().map(|t| left_tables.contains(t)).unwrap_or(false);
-                let left_is_right =
-                    left_tbl_lower.as_ref().map(|t| t == &right_table_lower).unwrap_or(true);
-
-                if right_is_left && left_is_right && equi_join.is_none() {
-                    equi_join = Some(EquiJoinInfo {
-                        left_col: right_col.to_string(),
-                        right_col: left_col.to_string(),
-                    });
-                    continue;
-                }
-            }
-        }
-
-        // Check if this predicate references only the right table
-        // (We'll add it to right_only_preds)
-        right_only_preds.push(pred);
-    }
-
-    equi_join.map(|ej| {
-        let filter = combine_with_and(right_only_preds);
-        (ej, filter)
-    })
-}
-
-/// Slot in a lookup key template
-#[derive(Debug, Clone)]
-enum KeySlot {
-    Constant(vibesql_types::SqlValue),
-    JoinKey,
-}
-
-/// Extract constant values from right filters that match PK columns.
-/// Returns a map from column name to constant value.
-fn extract_constant_prefix_for_pk(
-    right_filters: &Option<vibesql_ast::Expression>,
-    pk_columns: &[String],
-    join_key_col: &str,
-) -> HashMap<String, vibesql_types::SqlValue> {
-    let mut constants: HashMap<String, vibesql_types::SqlValue> = HashMap::new();
-
-    let filter = match right_filters {
-        Some(f) => f,
-        None => return constants,
-    };
-
-    // Flatten and look for equality predicates with constants
-    let conjuncts = flatten_conjuncts(filter);
-    let join_key_upper = join_key_col.to_lowercase();
-
-    for pred in conjuncts {
-        if let vibesql_ast::Expression::BinaryOp {
-            left,
-            op: vibesql_ast::BinaryOperator::Equal,
-            right,
-        } = &pred
-        {
-            // Check if left is column and right is literal
-            if let (
-                vibesql_ast::Expression::ColumnRef(col_id),
-                vibesql_ast::Expression::Literal(value),
-            ) = (left.as_ref(), right.as_ref())
-            {
-                let column = col_id.column_canonical();
-                let col_upper = column.to_lowercase();
-                // Only add if it's a PK column and not the join key
-                if pk_columns.iter().any(|pk| pk.to_lowercase() == col_upper)
-                    && col_upper != join_key_upper
-                {
-                    constants.insert(col_upper, value.clone());
-                }
-            }
-
-            // Check reverse: right is column and left is literal
-            if let (
-                vibesql_ast::Expression::Literal(value),
-                vibesql_ast::Expression::ColumnRef(col_id),
-            ) = (left.as_ref(), right.as_ref())
-            {
-                let column = col_id.column_canonical();
-                let col_upper = column.to_lowercase();
-                if pk_columns.iter().any(|pk| pk.to_lowercase() == col_upper)
-                    && col_upper != join_key_upper
-                {
-                    constants.insert(col_upper, value.clone());
-                }
-            }
-        }
-    }
-
-    constants
-}
-
-/// Build a lookup key template based on PK columns.
-fn build_lookup_key_template(
-    pk_columns: &[String],
-    join_key_col: &str,
-    constant_prefix: &HashMap<String, vibesql_types::SqlValue>,
-) -> Option<Vec<KeySlot>> {
-    let join_key_upper = join_key_col.to_lowercase();
-    let mut template = Vec::with_capacity(pk_columns.len());
-
-    for pk_col in pk_columns {
-        let pk_upper = pk_col.to_lowercase();
-
-        if pk_upper == join_key_upper {
-            template.push(KeySlot::JoinKey);
-        } else if let Some(value) = constant_prefix.get(&pk_upper) {
-            template.push(KeySlot::Constant(value.clone()));
-        } else {
-            // Missing a PK component, can't use INL
-            return None;
-        }
-    }
-
-    Some(template)
-}
-
-/// Build residual filter (predicates not covered by PK lookup)
-fn build_residual_filter(
-    right_filters: &Option<vibesql_ast::Expression>,
-    pk_columns: &[String],
-) -> Option<vibesql_ast::Expression> {
-    let filter = match right_filters {
-        Some(f) => f,
-        None => return None,
-    };
-
-    let pk_upper: HashSet<String> = pk_columns.iter().map(|s| s.to_lowercase()).collect();
-    let conjuncts = flatten_conjuncts(filter);
-
-    // Keep predicates that are NOT equality on PK columns
-    let residual: Vec<vibesql_ast::Expression> = conjuncts
-        .into_iter()
-        .filter(|pred| {
-            // Check if this is an equality predicate on a PK column
-            if let vibesql_ast::Expression::BinaryOp {
-                left,
-                op: vibesql_ast::BinaryOperator::Equal,
-                right,
-            } = pred
-            {
-                // Check if it's col = literal or literal = col
-                if let vibesql_ast::Expression::ColumnRef(col_id) = left.as_ref() {
-                    if pk_upper.contains(&col_id.column_canonical().to_lowercase())
-                        && matches!(right.as_ref(), vibesql_ast::Expression::Literal(_))
-                    {
-                        return false; // Filter out, covered by PK lookup
-                    }
-                }
-                if let vibesql_ast::Expression::ColumnRef(col_id) = right.as_ref() {
-                    if pk_upper.contains(&col_id.column_canonical().to_lowercase())
-                        && matches!(left.as_ref(), vibesql_ast::Expression::Literal(_))
-                    {
-                        return false; // Filter out, covered by PK lookup
-                    }
-                }
-            }
-            true // Keep this predicate
-        })
-        .collect();
-
-    combine_with_and(residual)
-}
-
-/// Find the index of a column in the schema
-fn find_column_index(schema: &CombinedSchema, col_name: &str) -> Option<usize> {
-    let col_upper = col_name.to_lowercase();
-    let mut offset = 0;
-
-    for (_, table_schema) in schema.table_schemas.values() {
-        for (idx, col) in table_schema.columns.iter().enumerate() {
-            if col.name.to_lowercase() == col_upper {
-                return Some(offset + idx);
-            }
-        }
-        offset += table_schema.columns.len();
-    }
-
-    None
-}
-
-/// Prefix-scan based semi-join for cases where an index exists but not all key columns
-/// are covered by the join condition.
-///
-/// This handles cases like TPC-H Q4:
-/// - Join on `o_orderkey = l_orderkey`
-/// - Index on `(l_orderkey, l_linenumber)`
-/// - Additional filter: `l_commitdate < l_receiptdate`
-///
-/// Instead of building a hash table from 60K lineitem rows, we do ~500 prefix scans
-/// (one per qualifying order), which is much faster when the left side is small.
-fn try_prefix_scan_semi_join(
-    left_result: &super::FromResult,
-    right_from: &vibesql_ast::FromClause,
-    condition: &Option<vibesql_ast::Expression>,
-    database: &vibesql_storage::Database,
-) -> Result<Option<super::FromResult>, ExecutorError> {
-    let debug = std::env::var("PREFIX_SEMI_DEBUG").is_ok();
-
-    // Must have a join condition
-    let cond = match condition {
-        Some(c) => c,
-        None => return Ok(None),
-    };
-
-    // Right side must be a simple table (not a join or subquery)
-    let (right_table_name, _right_alias) = match right_from {
-        vibesql_ast::FromClause::Table { name, alias, .. } => (name.clone(), alias.clone()),
-        _ => return Ok(None),
-    };
-
-    // Get the right table
-    let right_table = match database.get_table(&right_table_name) {
-        Some(t) => t,
-        None => return Ok(None),
-    };
-
-    // Parse the join condition to extract equi-join columns and additional filters
-    let (equi_join, right_filters) =
-        match parse_semi_join_condition(cond, left_result, &right_table_name) {
-            Some(parsed) => parsed,
-            None => return Ok(None),
-        };
-
-    if debug {
-        eprintln!(
-            "[PREFIX_SEMI] left_rows={}, join on {}={}, right_filters={:?}",
-            left_result.as_slice().len(),
-            equi_join.left_col,
-            equi_join.right_col,
-            right_filters
-        );
-    }
-
-    // Find an index on the right table that starts with the join key column
-    let right_col_upper = equi_join.right_col.to_lowercase();
-    let index_names = database.list_indexes_for_table(&right_table_name);
-
-    // Find an index where the first column is the join key
-    let mut usable_index_name: Option<String> = None;
-    for index_name in &index_names {
-        if let Some(idx_metadata) = database.get_index(index_name) {
-            if let Some(first_col) = idx_metadata.columns.first() {
-                if first_col.expect_column_name().to_lowercase() == right_col_upper {
-                    usable_index_name = Some(index_name.clone());
-                    break;
-                }
-            }
-        }
-    }
-
-    let index_name = match usable_index_name {
-        Some(name) => name,
-        None => {
-            if debug {
-                eprintln!(
-                    "[PREFIX_SEMI] No index starting with {} on {}",
-                    equi_join.right_col, right_table_name
-                );
-            }
-            return Ok(None);
-        }
-    };
-
-    let index = match database.get_index_data(&index_name) {
-        Some(idx) => idx,
-        None => return Ok(None),
-    };
-
-    if debug {
-        eprintln!(
-            "[PREFIX_SEMI] Using index {} for prefix scan on {}",
-            index_name, right_table_name
-        );
-    }
-
-    // Get left column index for extracting join keys
-    let left_col_idx = match find_column_index(&left_result.schema, &equi_join.left_col) {
-        Some(idx) => idx,
-        None => return Ok(None),
-    };
-
-    // Build evaluator for right-side filter (e.g., l_commitdate < l_receiptdate)
-    let right_schema =
-        CombinedSchema::from_table(right_table_name.clone(), right_table.schema.clone());
-    let evaluator = right_filters
-        .as_ref()
-        .map(|_| CombinedExpressionEvaluator::with_database(&right_schema, database));
-
-    // Collect distinct join keys from left side
-    let left_slice = left_result.as_slice();
-    let mut seen_keys: AHashSet<vibesql_types::SqlValue> =
-        AHashSet::with_capacity(left_slice.len());
-    let mut matching_keys: AHashSet<vibesql_types::SqlValue> = AHashSet::new();
-
-    let start = std::time::Instant::now();
-
-    for left_row in left_slice {
-        let join_key = &left_row.values[left_col_idx];
-
-        // Skip NULL join keys
-        if *join_key == vibesql_types::SqlValue::Null {
-            continue;
-        }
-
-        // Skip duplicate keys (we only need one match for SEMI join)
-        if seen_keys.contains(join_key) {
-            continue;
-        }
-        seen_keys.insert(join_key.clone());
-
-        // Do prefix lookup: find all rows where the first index column = join_key
-        let row_indices = index.prefix_multi_lookup(std::slice::from_ref(join_key));
-
-        // Check if any of the matching rows pass the additional filter
-        for row_idx in row_indices {
-            // Get the row data
-            let right_row = match right_table.get_row(row_idx) {
-                Some(row) => row,
-                None => continue, // Row deleted or invalid
-            };
-
-            // Apply filter if any
-            let passes = if let (Some(filter), Some(eval)) = (&right_filters, &evaluator) {
-                eval.clear_cse_cache();
-                matches!(eval.eval(filter, right_row), Ok(vibesql_types::SqlValue::Boolean(true)))
-            } else {
-                true // No filter
-            };
-
-            if passes {
-                matching_keys.insert(join_key.clone());
-                break; // Semi-join: we only need one match per left row
-            }
-        }
-    }
-
-    if debug {
-        let elapsed = start.elapsed();
-        eprintln!(
-            "[PREFIX_SEMI] {} distinct keys, {} matches, {:?} elapsed",
-            seen_keys.len(),
-            matching_keys.len(),
-            elapsed
-        );
-    }
-
-    // Build result: all left rows whose join key is in matching_keys
-    let result_rows: Vec<vibesql_storage::Row> = left_slice
-        .iter()
-        .filter(|row| {
-            let key = &row.values[left_col_idx];
-            matching_keys.contains(key)
-        })
-        .cloned()
-        .collect();
-
-    Ok(Some(super::FromResult::from_rows(left_result.schema.clone(), result_rows)))
 }
 
 /// Generate a join condition from USING columns with proper table qualification.
